Service config

+++ atlasdb-client/src/main/java/com/palantir/atlasdb/qos/AtlasDbQosClient.java
@@ -26,7 +26,7 @@ import javax.naming.LimitExceededException;
 public class AtlasDbQosClient {
     QosService qosService;
 
-    volatile int credits;
+    volatile long credits;
 
     private AtlasDbQosClient(QosService qosService) {
         this.qosService = qosService;
